# fullstack-projects/jobs/backend/swagger.yaml
openapi: 3.0.0

info:
  title: Jobs API

  contact: {}

  version: "1.0"

servers:
  - url: https://expressjs-mongoose-production-40dd.up.railway.app/api/v1

    variables: {}

paths:
  /auth/register:
    post:
      tags:
        - Auth

      summary: Register

      operationId: Register

      parameters: []

      requestBody:
        description: ""

        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/RegisterRequest"

                - example:
                    name: vahe

                    email: vahe.aa1@yandex.com

                    password: 12bb12bb

            example:
              name: vahe

              email: vahe.aa1@yandex.com

              password: 12bb12bb

        required: true

      responses:
        "200":
          description: ""

          headers: {}

      deprecated: false

      security: []

  /auth/login:
    post:
      tags:
        - Auth

      summary: Login

      operationId: Login

      parameters: []

      requestBody:
        description: ""

        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/LoginRequest"

                - example:
                    email: vahe.aa1@yandex.com

                    password: 12bb12bb

            example:
              email: vahe.aa1@yandex.com

              password: 12bb12bb

        required: true

      responses:
        "200":
          description: ""

          headers: {}

      deprecated: false

      security: []

  /jobs:
    post:
      tags:
        - Jobs

      summary: Create New Job

      operationId: CreateNewJob

      parameters: []

      requestBody:
        description: ""

        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CreateNewJobRequest"

                - example:
                    company: Facebook

                    position: Moderator

            example:
              company: Facebook

              position: Moderator

        required: true

      responses:
        "200":
          description: ""

          headers: {}

      deprecated: false

      security:
        - bearer: []

    get:
      tags:
        - Jobs

      summary: Get All Jobs

      operationId: GetAllJobs

      parameters: []

      responses:
        "200":
          description: ""

          headers: {}

      deprecated: false

      security:
        - bearer: []

  /jobs/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The job ID
    get:
      tags:
        - Jobs

      summary: Get Single Job

      operationId: GetSingleJob

      parameters: []

      responses:
        "200":
          description: ""

          headers: {}

      deprecated: false

      security:
        - bearer: []

  /jobs/{jobId}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The job ID
    patch:
      tags:
        - Jobs

      summary: Update Job

      operationId: UpdateJob

      parameters: []

      requestBody:
        description: ""

        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/UpdateJobRequest"

                - example:
                    company: Netflix

                    position: SMM specialist

            example:
              company: Netflix

              position: SMM specialist

        required: true

      responses:
        "200":
          description: ""

          headers: {}

      deprecated: false

      security:
        - bearer: []

    delete:
      tags:
        - Jobs

      summary: Delete Job

      operationId: DeleteJob

      parameters: []

      responses:
        "200":
          description: ""

          headers: {}

      deprecated: false

      security:
        - bearer: []

components:
  schemas:
    RegisterRequest:
      title: RegisterRequest

      required:
        - name

        - email

        - password

      type: object

      properties:
        name:
          type: string

        email:
          type: string

        password:
          type: string

      example:
        name: vahe

        email: vahe.aa1@yandex.com

        password: 12bb12bb

    LoginRequest:
      title: LoginRequest

      required:
        - email

        - password

      type: object

      properties:
        email:
          type: string

        password:
          type: string

      example:
        email: vahe.aa1@yandex.com

        password: 12bb12bb

    CreateNewJobRequest:
      title: CreateNewJobRequest

      required:
        - company

        - position

      type: object

      properties:
        company:
          type: string

        position:
          type: string

      example:
        company: Facebook

        position: Moderator

    UpdateJobRequest:
      title: UpdateJobRequest

      required:
        - company

        - position

      type: object

      properties:
        company:
          type: string

        position:
          type: string

      example:
        company: Netflix

        position: SMM specialist

  securitySchemes:
    bearer:
      type: http

      scheme: bearer

security: []

tags:
  - name: Auth

  - name: Jobs